Add ThreadedSocket
Starts a thread, and uses it to send messages to the core agent out of
band from the web request being handled.

The thread is marked as a `daemon` thread, which means that it won't
hold up the exit of the program. But that means that any messages in the
Queue when the app shuts down will be dropped, and not finished. This is
highly unlikely to be a meaningful situation, since we already have the
situation of being `kill -9`d by the user, or the managing web server.
So we need to be sure to handle that cleanly anyway.