Default a None timestamp in any Command to now
This can come up if for some reason a Span isn't stopped, and its
timestamp is None. The code to create the StopSpan command explicitly
sets the kwarg to None, rather than leaving it off. So the defaulting
in `get` that existed didn't work (it was there as a key, just with a
value of None).

I applied this change to all Commands, since timestamps should always
exist, even if off.