Default a None timestamp in any Command (#52)
* Default a None timestamp in any Command to now

This can come up if for some reason a Span isn't stopped, and its
timestamp is None. The code to create the StopSpan command explicitly
sets the kwarg to None, rather than leaving it off. So the defaulting
in `get` that existed didn't work (it was there as a key, just with a
value of None).

I applied this change to all Commands, since timestamps should always
exist, even if off.

* Avoid creating invalid BatchCommands

* Create an INVALID_DATE constant to fill in and prevent JSON errors

This is a fallback, and not for normal error cases. The function
from_tracked_request will not create None timestamps in StopSpan or
FinishRequest commands. But if for some reason a command manages to be
created without a timestamp, this will prevent it from exceptioning when
we serialize it.

* Fix mistaken call to datetime

* Add tests for StopSpan, FinishRequest and BatchCommand for unstopped requests