Add additional thread locking around CoreAgentSocket
We were running into an issue (reproducable on 2.7, although likely
existing in other versions) where a socket thread was created, but then
not stored in the _instance variable. So it held the run_lock, but
wasn't available to be looked up.

This change adds a new lock around the instance call, which will prevent
race conditions around looking up, creating and recreating the socket
thread.

Also, it shifts the run_lock acquisition to the calling thread, instead
of being the first thing done in the newly created thread.