AgentContext is a global, not thread-local
We want all AgentContext references to point to the same set of internal
data (configs for now, and whatever we add going down the line). Before,
it was a thread local, so each copy had a distinct instance of the
ScoutConfig class, which would make editing the config in the running
system much harder, and generally is surprising if you expect config to
be 'global'.