Always read the response to CoreAgent messages
This appears to fix the socket timeout errors in #125.

Previously we were simply ignoring the contents of any response message
from the core agent, not even reading off the socket. We were seeing
`timed out` exceptions. By reading off the socket every time, and fixing
the `recv` calls to actually read *all* of the data off the socket, the
timeouts go away on my system.