Enable __future__ features consistently.
Include all features that are still optional in Python 2.7 and become
mandatory in Python 3.0.

Special case: remove the __future__ features from
scout_apm/core/monkey.py which is third-party code.
absolute_import used to be enabled but it had no effect
because that module didn't contain any import statements.

The correctness of this change can be verified as follows:

    # All non-empty Python files contain a __future__ import
    # (except scout_apm/core/monkey.py, as discussed above,
    # and the Sphinx configuration which doesn't matter here.)

    $ find . -name '*.py' ! -size 0 -print0 | xargs -0 grep -L __future__
    ./docs/conf.py
    ./src/scout_apm/core/monkey.py

    # The __future__ import is consistent across all these files.

    $ find . -name '*.py' ! -size 0 -print0 | xargs -0 grep -h __future__ | uniq
    from __future__ import absolute_import, division, print_function, unicode_literals