Improve and simplify Travis CI setup. (#135)
* Run tests against built wheels.

Previously, CIBW_TEST_COMMAND called tox, which created new virtualenvs
for running test. This defeated the purpose of CIBW_TEST_COMMAND, namely
testing with the result of installing the freshly-built wheel.

Pick the version of Djang that exercises the most code paths. It doesn't
seem worthwhile to re-run tests against multiple versions of Djano here.

* Refactor conditions that are always true.

* Move definition of env variables to global section.

This removes a difference with the recommended configuration of
cibuildwheel.

Also remove CIBW_BUILD_VERBOSITY which was set to its default value.

* Configure according to the cibuildwheel docs.

* Keep the default verbosity of py.test.

This avoids a massive output when running tox.

* Run tox on all supported Python versions.

* Introduce a Python version matrix so that tests run on
  all supported Python version in parallel.
* Separate "test" and "build" stages so that builds run only
  when tests pass on all Python versions.

* Declare support for Python 2.7 in setup.py.

This isn't true yet but makes it possible to run tests with tox.

Remove __future__ imports from setup.py because setuptools dislikes
unicode and none of the other features are useful in this script.

* Run tests with Python 2.7 on Travis CI.

* Run style checks on Travis CI.

* Use the default target locations.

* Remove unnecessary entry in MANIFEST.in.