Remove all references to PyYAML (#146)
A user reported that an automated GitHub security scan reported PyYAML
as being vulnerable [1].  The user also said that it doesn't appear that
we even use PyYAML. After some checking, it looks that he's right, we
don't use it.

My memory of how this dependency was ever added is that we pulled it in
as a dependency very early in development to implement configuration in
a similar way to how the Ruby agent does. But then, we implemented a
more Pythonic system for configuration.

The underlying security issue isn't relevant to or exploitable in our
agent. We never even import the PyYAML into our code, much less use the
potentially vulnerable `load` function. But of course, the most secure
code is code that doesn't exist. So this commit removes all mentions of
'pyyaml'.

[Closes #145]

[1]: https://github.com/scoutapp/scout_apm_python/issues/145