Make Django middleware addition idempotent
Whilst improving the tests, I discovered that the way the Django test setup works, the app.ready() gets called in every test. This was causing the middleware to have multiple copies inserted.

Change this so that if a middleware is already in the stack, it doesn't get re-added. This will also let users organize the middleware if they so wish.