Refactor Django configuration (#170)
The existing tests did not reset Django to a clean state each time, since the scout_apm app was being repeatedly re-installed and uninstalled, but it didn't do anything to reset settings on uninstall. This is what lead to me implementing idempotent middleware addition in #168.

This PR refactors the way configuration is done. The test Django app is initialized only once, with Scout already installed. Scout then synchronizes the Django settings into ScoutConfig, and listens to Django's `setting_changed` signal to keep them in sync. Rather than rely on reset_all at the end of each test, I added an assertion that no settings have been left in place by any test - if they all use `override_settings` to change the settings, this will be the case.

This has let me write tests that cover all the cases with the middleware idempotency.