Capture query string in path tag (#268)
First step in implementing #249.

Add a funtion `create_filtered_path`. This combines the path and filtered query parameters. Params are filtered according to a list of banned strings I collected from Ruby on Rails (as per the Ruby agent) and extras from Sentry.

The implementation of query parameters varies by framework so a little custom iteration code was needed on each.

Next step is to add a config value to allow disabling the capturing, back to the previous behaviour only capturing the path.