Remove overly cautious header fetching behaviour
Both the Bottle and Pyramid integrations have this overly cautious behaviour around reading headers from the upstream framework. They both assume that basic framework operations such as fetching a header can fail, and guard around it.

Some of this came from my refactoring for Bottle in #192 and Pyramid in #195, where I narrowed large `try/except` blocks down, but didn't rethink this paranoia.

I think this is too cautious and can be removed. It required a lot of extra lines for the two frameworks when adding Amazon request queue timing in #279, to really no extra effect. If fetching headers (with safe defaults) is broken, users' applications are pretty much guaranteed to be broken too.