Make config object global (#322)
Whilst investigating making the agent context configurable with running under async, I realized the `ScoutConfig` class does little as an instance - all the state is stored in globals and classes. Yet we recreate instances all over the place.

This PR turns into a true global object, a single instance of the class, so it's easier to import and share.