Revert utcnow change
This has lead to timezone offsets being sent to the Core Agent, which it doesn't parse, logging this error and failing to return a full response to the client:

```
[2019-11-25T18:09:49][socket::server][INFO] Reading a command from stream: NoParse(Error("trailing input", line: 1, column: 70), "{\"ApplicationEvent\": {\"timestamp\": \"2019-11-25T18:09:49.328081+00:00Z\", \"event_type\": \"CPU/Utilization\", \"event_value\": 0.027687317280944007, \"source\": \"Pid: 69721\"}}")
```

I think the safest thing to do here is a pure revert. It would be better theoretically to track datetimes with timezone information but it seems too disruptive to the way the core agent works.