Remove tracking queue time from amazon trace id header.
The header's time value is in seconds which isn't enough precision
for us to provide quality data to users. With seconds precision, the
queue time hovers about 0.5s due to the truncation.

It may be possible to pull the millisecond precision data from the
AWS logs themselves and avoid the python agent altogether.

Closes #631