Monitor errors via got_request_exception signal.
The middleware process_exception would only allow us to track
exceptions thrown in views, that weren't picked up by an
earlier middleware that returned a result. This solution allows
us to track exceptions that occur anywhere in the application.

The downside to this is that it's possible if another receiver
of got_request_exception throws a new exception, our handler
may get the wrong exception with sys.exc_info().

We could limit the scope of that issue by injecting a middleware
that's as far down the list as possible to wrap the view
and catch exceptions there. Then rely on the signal handler
for catching exceptions in the middleware.

Create ErrorServiceThread to buffer error messages.

Implement buffering logic, write integration tests for error_service.

Exclude specific libraries' logs from caplog.

Some of our checks are positional based, but have the chance of triggering
a libraries own debug statements. For cases we know we can ignore,
we can add them to the filter to make the tests more robust.

Expose a configurable API and ErrorMonitor interface.

Support Celery for the error monitoring solution.

Custom params belong on the context.

Celery's custom params go into a celery key of custom params.
Custom controller overrides the request_component's controller.